final update to dual-language for other files left

+++ roots.c
@@ -52,21 +52,21 @@ void load_volume_table() {
 
     fstab = fs_mgr_read_fstab("/etc/recovery.fstab");
     if (!fstab) {
-#ifndef USE_CHINESE_FONT
+if ( language== 1 )
         LOGE("failed to read /etc/recovery.fstab\n");
-#else
+else
         LOGE("读取 /etc/recovery.fstab 失败\n");
-#endif
+
         return;
     }
 
     ret = fs_mgr_add_entry(fstab, "/tmp", "ramdisk", "ramdisk", 0);
     if (ret < 0 ) {
-#ifndef USE_CHINESE_FONT
+if ( language== 1 )
         LOGE("failed to add /tmp entry to fstab\n");
-#else
+else
         LOGE("添加条目 /tmp 到 fstab 中失败\n");
-#endif
+
         fs_mgr_free_fstab(fstab);
         fstab = NULL;
         return;
@@ -89,11 +89,11 @@ void load_volume_table() {
     device_truedualboot_after_load_volume_table();
 #endif
 
-#ifndef USE_CHINESE_FONT
+if ( language== 1 )
     fprintf(stderr, "recovery filesystem table\n");
-#else
+else
     fprintf(stderr, "   recovery 文件系统表\n");
-#endif
+
     fprintf(stderr, "=========================\n");
     for (i = 0; i < fstab->num_entries; ++i) {
         Volume* v = &fstab->recs[i];
@@ -183,11 +183,11 @@ int try_mount(const char* device, const char* mount_point, const char* fs_type, 
     }
     if (ret == 0)
         return 0;
-#ifndef USE_CHINESE_FONT
+if ( language== 1 )
     LOGW("failed to mount %s (%s)\n", device, strerror(errno));
-#else
+else
     LOGW("挂载 %s 失败(%s)\n", device, strerror(errno));
-#endif
+
     return ret;
 }
 
@@ -248,11 +248,11 @@ int ensure_path_mounted_at_mount_point(const char* path, const char* mount_point
 
     if (is_data_media_volume_path(path)) {
         if (ui_should_log_stdout()) {
-#ifndef USE_CHINESE_FONT
+if ( language== 1 )
             LOGI("using /data/media for %s.\n", path);
-#else
+else
             LOGI("为 %s 使用 /data/media。\n", path);
-#endif
+
         }
         int ret;
         if (0 != (ret = ensure_path_mounted("/data")))
@@ -262,11 +262,11 @@ int ensure_path_mounted_at_mount_point(const char* path, const char* mount_point
     }
     Volume* v = volume_for_path(path);
     if (v == NULL) {
-#ifndef USE_CHINESE_FONT
+if ( language== 1 )
         LOGE("unknown volume for path [%s]\n", path);
-#else
+else
         LOGE("未知的卷路径 [%s]\n", path);
-#endif
+
         return -1;
     }
     if (strcmp(v->fs_type, "ramdisk") == 0) {
@@ -277,11 +277,11 @@ int ensure_path_mounted_at_mount_point(const char* path, const char* mount_point
     int result;
     result = scan_mounted_volumes();
     if (result < 0) {
-#ifndef USE_CHINESE_FONT
+if ( language== 1 )
         LOGE("failed to scan mounted volumes\n");
-#else
+else
         LOGE("扫描已挂载的卷失败\n");
-#endif
+
         return -1;
     }
 
@@ -306,11 +306,11 @@ int ensure_path_mounted_at_mount_point(const char* path, const char* mount_point
         const MtdPartition* partition;
         partition = mtd_find_partition_by_name(v->blk_device);
         if (partition == NULL) {
-#ifndef USE_CHINESE_FONT
-            LOGE("failed to find \"%s\" partition to mount at \"%s\"\n",
-#else
+if ( language== 1 )
+            LOGE("failed to find \"%s\" partition to mount at \"%s\"\n",v->blk_device, mount_point);
+else
             LOGE("查找分区 \"%s\" 并挂载到 \"%s\" 失败\n",
-#endif
+
                  v->blk_device, mount_point);
             return -1;
         }
@@ -352,11 +352,11 @@ int ensure_path_unmounted(const char* path) {
 
     Volume* v = volume_for_path(path);
     if (v == NULL) {
-#ifndef USE_CHINESE_FONT
+if ( language== 1 )
         LOGE("unknown volume for path [%s]\n", path);
-#else
+else
         LOGE("未知的卷路径 [%s]\n", path);
-#endif
+
         return -1;
     }
 
@@ -368,11 +368,11 @@ int ensure_path_unmounted(const char* path) {
     int result;
     result = scan_mounted_volumes();
     if (result < 0) {
-#ifndef USE_CHINESE_FONT
+if ( language== 1 )
         LOGE("failed to scan mounted volumes\n");
-#else
+else
         LOGE("扫描已挂载的卷失败\n");
-#endif
+
         return -1;
     }
 
@@ -410,22 +410,22 @@ int format_volume(const char* volume) {
     if (v == NULL) {
         // silent failure for sd-ext
         if (strcmp(volume, "/sd-ext") != 0)
-#ifndef USE_CHINESE_FONT
+if ( language== 1 )
             LOGE("unknown volume '%s'\n", volume);
-#else
+else
             LOGE("未知卷 '%s\'\n", volume);
-#endif
+
         return -1;
     }
     // silent failure to format non existing sd-ext when defined in recovery.fstab
     if (strcmp(volume, "/sd-ext") == 0) {
         struct stat s;
         if (0 != stat(v->blk_device, &s)) {
-#ifndef USE_CHINESE_FONT
+if ( language== 1 )
             LOGI("Skipping format of sd-ext\n");
-#else
+else
             LOGI("跳过对 sd-ext 的格式化\n");
-#endif
+
             return -1;
         }
     }
@@ -434,42 +434,42 @@ int format_volume(const char* volume) {
     // formatted instead of /storage/sdcard0/.android_secure
     if (fs_mgr_is_voldmanaged(v) && strcmp(volume, v->mount_point) == 0) {
         if (ensure_path_unmounted(volume) != 0) {
-#ifndef USE_CHINESE_FONT
+if ( language== 1 )
             LOGE("format_volume failed to unmount %s", v->mount_point);
-#else
+else
             LOGE("format_volume 卸载 \"%s\" 时出错\n", v->mount_point);
-#endif
+
         }
         return vold_format_volume(v->mount_point, 1) == CommandOkay ? 0 : -1;
     }
 
     if (strcmp(v->fs_type, "ramdisk") == 0) {
         // you can't format the ramdisk.
-#ifndef USE_CHINESE_FONT
+if ( language== 1 )
         LOGE("can't format_volume \"%s\"", volume);
-#else
+else
         LOGE("无法格式化卷 \"%s\"", volume);
-#endif
+
         return -1;
     }
     if (strcmp(v->mount_point, volume) != 0) {
 #if 0
-#ifndef USE_CHINESE_FONT
+if ( language== 1 )
         LOGE("can't give path \"%s\" to format_volume\n", volume);
-#else
+else
         LOGE("无法传递路径 \"%s\" 给 format_volume\n", volume);
-#endif
+
         return -1;
 #endif
         return format_unknown_device(v->blk_device, volume, NULL);
     }
 
     if (ensure_path_unmounted(volume) != 0) {
-#ifndef USE_CHINESE_FONT
+if ( language== 1 )
         LOGE("format_volume failed to unmount \"%s\"\n", v->mount_point);
-#else
+else
         LOGE("format_volume 卸载 \"%s\" 时出错\n", v->mount_point);
-#endif
+
         return -1;
     }
 
@@ -477,36 +477,36 @@ int format_volume(const char* volume) {
         mtd_scan_partitions();
         const MtdPartition* partition = mtd_find_partition_by_name(v->blk_device);
         if (partition == NULL) {
-#ifndef USE_CHINESE_FONT
+if ( language== 1 )
             LOGE("format_volume: no MTD partition \"%s\"\n", v->blk_device);
-#else
+else
             LOGE("format_volume: 无 MTD 分区 \"%s\"\n", v->blk_device);
-#endif
+
             return -1;
         }
 
         MtdWriteContext *write = mtd_write_partition(partition);
         if (write == NULL) {
-#ifndef USE_CHINESE_FONT
+if ( language== 1 )
             LOGW("format_volume: can't open MTD \"%s\"\n", v->blk_device);
-#else
+else
             LOGW("format_volume: 无法打开 MTD 设备 \"%s\"\n", v->blk_device);
-#endif
+
             return -1;
         } else if (mtd_erase_blocks(write, -1) == (off_t) -1) {
-#ifndef USE_CHINESE_FONT
+if ( language== 1 )
             LOGW("format_volume: can't erase MTD \"%s\"\n", v->blk_device);
-#else
+else
             LOGW("format_volume: 无法擦除 MTD 设备 \"%s\"\n", v->blk_device);
-#endif
+
             mtd_write_close(write);
             return -1;
         } else if (mtd_write_close(write)) {
-#ifndef USE_CHINESE_FONT
+if ( language== 1 )
             LOGW("format_volume: can't close MTD \"%s\"\n", v->blk_device);
-#else
+else
             LOGW("format_volume: 无法关闭 MTD 设备 \"%s\"\n", v->blk_device);
-#endif
+
             return -1;
         }
         return 0;
@@ -515,11 +515,11 @@ int format_volume(const char* volume) {
     if (strcmp(v->fs_type, "ext4") == 0) {
         int result = make_ext4fs(v->blk_device, v->length, volume, sehandle);
         if (result != 0) {
-#ifndef USE_CHINESE_FONT
+if ( language== 1 )
             LOGE("format_volume: make_extf4fs failed on %s\n", v->blk_device);
-#else
+else
             LOGE("format_volume: 在设备 %s 上执行 make_extf4fs 时出错\n", v->blk_device);
-#endif
+
             return -1;
         }
         return 0;
@@ -529,11 +529,11 @@ int format_volume(const char* volume) {
     if (strcmp(v->fs_type, "f2fs") == 0) {
         char* args[] = { "mkfs.f2fs", v->blk_device };
         if (make_f2fs_main(2, args) != 0) {
-#ifndef USE_CHINESE_FONT
+if ( language== 1 )
             LOGE("format_volume: mkfs.f2fs failed on %s\n", v->blk_device);
-#else
+else
             LOGE("format_volume: 对设备 %s 执行 mkfs.f2fs 失败\n", v->blk_device);
-#endif
+
             return -1;
         }
         return 0;
@@ -541,11 +541,11 @@ int format_volume(const char* volume) {
 #endif
 
 #if 0
-#ifndef USE_CHINESE_FONT
+if ( language== 1 )
     LOGE("format_volume: fs_type \"%s\" unsupported\n", v->fs_type);
-#else
+else
     LOGE("format_volume: fs_type \"%s\" 不支持\n", v->fs_type);
-#endif
+
     return -1;
 #endif
     return format_unknown_device(v->blk_device, volume, v->fs_type);
